Use useReducer and move keydown to window
Removed the need for Input tags so that we don't need to keep track of
inputs all over the grid which created extra renders and event handlers.

indexCurrent was renamed to index for clarity and since only one index
is set in state at any moment.

useReducer was used to help separate state changes from components.